+++ src/index.ts
@@ -1,4 +1,5 @@
 export { usePlaidLink } from './usePlaidLink';
 export { PlaidLink } from './PlaidLink';
 export { PlaidEmbeddedLink } from './PlaidEmbeddedLink';
-export * from './types'
+export * from './types';
+export * from './sandbox-fix';